+++ cax/main.py
@@ -101,10 +101,14 @@ def main():
         corrections.AddElectronLifetime(),  # Add electron lifetime to run, which is just a function of calendar time
         corrections.AddGains(), #  Adds gains to a run, where this is computed using slow control information
         corrections.AddDriftVelocity(), #  Adds drift velocity to the run, also computed from slow control info
-        #corrections.AddSlowControlInformation(),  
+        corrections.SetS2xyMap(),
+        corrections.SetLightCollectionEfficiency(),
+        corrections.SetFieldDistortion(),
+        corrections.SetNeuralNetwork(),
+        #corrections.AddSlowControlInformation(),
         data_mover.CopyPush(),  # Upload data through e.g. scp or gridftp to this location where cax running
         #tsm_mover.AddTSMChecksum(), # Add forgotten Checksum for runDB for TSM client.
-	checksum.CompareChecksums(),  # See if local data corrupted
+        checksum.CompareChecksums(),  # See if local data corrupted
         clear.RetryStalledTransfer(),  # If data transferring e.g. 48 hours, probably cax crashed so delete then retry
         clear.RetryBadChecksumTransfer(),  # If bad checksum for local data and can fetch from somewhere else, delete our copy
 
@@ -254,17 +258,17 @@ def massive():
     while True: # yeah yeah
         query = {}
 
-        #t1 = datetime.datetime.utcnow()
-        #if t1 - t0 < dt:
-        #    logging.info("Iterative mode")
+        t1 = datetime.datetime.utcnow()
+        if t1 - t0 < dt:
+            logging.info("Iterative mode")
 
-        #    # See if there is something to do
-        #    query['start'] = {'$gt' : t0}
+            # See if there is something to do
+            query['start'] = {'$gt' : t0}
 
-        #    logging.info(query)
-        #else:
-        #    logging.info("Full mode")
-        #    t0 = t1
+            logging.info(query)
+        else:
+            logging.info("Full mode")
+            t0 = t1
 
 
         if args.run: